# storage/api-docs/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Admin application",
        "description": "API endpoints for the admin application.",
        "contact": {
            "name": "viracoding",
            "email": "viracodingGplus@gmail.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/admin",
            "description": "development Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Admin login and get Admin Data",
                "operationId": "a0472c8da28f5a65e86f7b8df183466b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "password"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "format": "phone number",
                                        "example": "1234567890"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "*****"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or failure to send SMS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "شما اجازه دسترسی ندارید ."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}